Netflix

+++ matrix_factorization/loss.cu
@@ -45,8 +45,8 @@ void calculate_loss_gpu(CudaDenseMatrix* P_d, CudaDenseMatrix* Q_d, config::Conf
 // and https://devblogs.nvidia.com/using-shared-memory-cuda-cc/
 // Fixes the problems related to data sizes
 template <unsigned int block_size>
-__global__ void total_loss_kernel(float *in_errors, float *out_errors, int n_errors, ErrorType error_type) {
-    extern __shared__ float sdata[];
+__global__ void total_loss_kernel(float *in_errors, double *out_errors, int n_errors, ErrorType error_type) {
+    extern __shared__ double sdata[];
     unsigned int tid = threadIdx.x;
     unsigned int i = blockIdx.x * block_size + tid;
     unsigned int grid_size = block_size * gridDim.x;
@@ -109,63 +109,63 @@ __global__ void total_loss_kernel(float *in_errors, float *out_errors, int n_err
 
 std::tuple<float, float> get_error_metrics_cpu(float *errors, float *errors_device, int n_errors) {
     cudaMemcpy(errors, errors_device, n_errors * sizeof(float), cudaMemcpyDeviceToHost);
-    float mae = 0.0;
-    float rmse = 0.0;
+    double mae = 0.0;
+    double rmse = 0.0;
     for(int k = 0; k <  n_errors; k++) {
         mae += abs(errors[k]);
         rmse += errors[k] * errors[k];
     }
     mae /= n_errors;
     rmse = sqrt(rmse / n_errors);
-    return std::make_tuple(mae, rmse);
+    return std::make_tuple((float)mae, (float)rmse);
 }
 
 // Inspired by https://developer.download.nvidia.com/assets/cuda/files/reduction.pdf
 // and https://devblogs.nvidia.com/using-shared-memory-cuda-cc/
 // Fixes the problems related to data sizes
-float calculate_error_metric_gpu(float *in_errors, float *out_errors, float *out_errors_host, int n_errors, int grid_size, int block_size, ErrorType error_type) {
+float calculate_error_metric_gpu(float *in_errors, double *out_errors, double *out_errors_host, int n_errors, int grid_size, int block_size, ErrorType error_type) {
     switch(block_size) {
         case 512:
-            total_loss_kernel<512><<<grid_size, block_size, 512 * sizeof(float)>>>(in_errors, out_errors, n_errors, error_type);
+            total_loss_kernel<512><<<grid_size, block_size, 512 * sizeof(double)>>>(in_errors, out_errors, n_errors, error_type);
             break;
         case 256:
-            total_loss_kernel<256><<<grid_size, block_size, 256 * sizeof(float)>>>(in_errors, out_errors, n_errors, error_type);
+            total_loss_kernel<256><<<grid_size, block_size, 256 * sizeof(double)>>>(in_errors, out_errors, n_errors, error_type);
             break;
         case 128:
-            total_loss_kernel<128><<<grid_size, block_size, 128 * sizeof(float)>>>(in_errors, out_errors, n_errors, error_type);
+            total_loss_kernel<128><<<grid_size, block_size, 128 * sizeof(double)>>>(in_errors, out_errors, n_errors, error_type);
             break;
         case 64:
-            total_loss_kernel< 64><<<grid_size, block_size,  64 * sizeof(float)>>>(in_errors, out_errors, n_errors, error_type);
+            total_loss_kernel< 64><<<grid_size, block_size,  64 * sizeof(double)>>>(in_errors, out_errors, n_errors, error_type);
             break;
         case 32:
-            total_loss_kernel< 32><<<grid_size, block_size,  32 * sizeof(float)>>>(in_errors, out_errors, n_errors, error_type);
+            total_loss_kernel< 32><<<grid_size, block_size,  32 * sizeof(double)>>>(in_errors, out_errors, n_errors, error_type);
             break;
         case 16:
-            total_loss_kernel< 16><<<grid_size, block_size,  16 * sizeof(float)>>>(in_errors, out_errors, n_errors, error_type);
+            total_loss_kernel< 16><<<grid_size, block_size,  16 * sizeof(double)>>>(in_errors, out_errors, n_errors, error_type);
             break;
         case 8:
-            total_loss_kernel<  8><<<grid_size, block_size,   8 * sizeof(float)>>>(in_errors, out_errors, n_errors, error_type);
+            total_loss_kernel<  8><<<grid_size, block_size,   8 * sizeof(double)>>>(in_errors, out_errors, n_errors, error_type);
             break;
         case 4:
-            total_loss_kernel<  4><<<grid_size, block_size,   4 * sizeof(float)>>>(in_errors, out_errors, n_errors, error_type);
+            total_loss_kernel<  4><<<grid_size, block_size,   4 * sizeof(double)>>>(in_errors, out_errors, n_errors, error_type);
             break;
         case 2:
-            total_loss_kernel<  2><<<grid_size, block_size,   2 * sizeof(float)>>>(in_errors, out_errors, n_errors, error_type);
+            total_loss_kernel<  2><<<grid_size, block_size,   2 * sizeof(double)>>>(in_errors, out_errors, n_errors, error_type);
             break;
         case 1:
-            total_loss_kernel<  1><<<grid_size, block_size,   1 * sizeof(float)>>>(in_errors, out_errors, n_errors, error_type);
+            total_loss_kernel<  1><<<grid_size, block_size,   1 * sizeof(double)>>>(in_errors, out_errors, n_errors, error_type);
             break;
     }
     CHECK_CUDA(cudaGetLastError());
-    CHECK_CUDA(cudaMemcpy(out_errors_host, out_errors, grid_size * sizeof(float), cudaMemcpyDeviceToHost));
-    float total = 0;
+    CHECK_CUDA(cudaMemcpy(out_errors_host, out_errors, grid_size * sizeof(double), cudaMemcpyDeviceToHost));
+    double total = 0;
     for(int k = 0; k < grid_size; k++) {
         total += out_errors_host[k];
     }
     return error_type == RMSE ? sqrt(total / n_errors) : total / n_errors;
 }
 
-std::tuple<float, float> get_error_metrics_gpu(float *in_errors, float *out_errors, float *out_errors_host, int n_errors, int grid_size, int block_size) {
+std::tuple<float, float> get_error_metrics_gpu(float *in_errors, double *out_errors, double *out_errors_host, int n_errors, int grid_size, int block_size) {
     float mae = calculate_error_metric_gpu(in_errors, out_errors, out_errors_host, n_errors, grid_size, block_size, MAE);
     float rmse = calculate_error_metric_gpu(in_errors, out_errors, out_errors_host, n_errors, grid_size, block_size, RMSE);
     return std::make_tuple(mae, rmse);
